Fix updating aws_sesv2_configuration_set to remove suppression reasons
This commit fixes the bug in issue #45526, so that updating
`suppressed_reasons` to an empty list for an existing configuration set
takes effect as expected.

I think the bug was being caused because PR #29671 which added the code
to allow the suppressed reasons list to be empty changed the method
`resourceConfigurationSetCreate` but not `resourceConfigurationSetUpdate`.

I copied the changes in commit 43b9eb9 to across to the code in
`resourceConfigurationSetUpdate` that reads `suppression_options`. I
don't fully understand the changes, but they seem to work... they make
the failing test `TestAccSESV2ConfigurationSet_suppressedReasonsAddEmpty`
added in the previous commit pass.